requirements.txt
Django
djangorestframework
celery
redis
psycopg[binary]
django-redis
faker
locust